Addresses #14: Add configuration for incoming attachment directory
As far as I'm aware, this is exclusively useful in the context of a Docker deployment where signald and mautrix-signal are in different containers AND the paths to their respective data directories are not identical.

This introduces a new configuration item, `incoming_attachment_dir`. When set and receiving/sending attachments, this translates a path from being mautrix-relative to being signald-relative. From an implementation perspective, this is on the "mautrix" side of things (portal.py). It could conceivably be on the "signald" side of things (signald.py) instead I found it was easier to modify the path on creation.

This enables easier integration with the `finn/signald` Docker image.