Automatically reconnect to signald
If the signald socket gets disconnected (e.g., signald restart), we will
now retry reconnecting to the socket forever, with a 5 second wait
between attempts.  We will also retry the initial connect forever if the
socket connection fails on startup.

This change creates a couple "synthetic" RPC events for connect and
disconnect that handlers can be attached to, as well as a NotConnected
exception class.  These should help callers deal with the fact that the
client may not always be connected.

Fixes #18