remove duplicate blurhash key
According to the blurhash proposal[0] "Implementations wishing to add
this before this MSC is merged can do so with the following: The
blurhash key in any events, request or response bodies should be
replaced with xyz.amorgan.blurhash."

The current implementation of mautrix-signal sets both blurhash and
xyz.amorgan.blurhash, which breaks for example the deserialization of
such events in Ruma 0.6.3. Ruma handles xyz.amorgan.blurhash as
blurhash, and when there's also a blurhash key present, it fails to
deserialize the event with "duplicate field" error.

This commit fixes the issue by just setting the xyz.amorgan.blurhash
key, as per the instructions in the proposal.

[0]: https://github.com/matrix-org/matrix-spec-proposals/blob/anoa/blurhash/proposals/2448-blurhash-for-media.md